+++ __tests__/utils/hash.test.ts
@@ -26,7 +26,7 @@ describe('Hash Tester', () => {
       const classHash = computeContractClassHash(compiledOpenZeppelinAccount);
 
       expect(classHash).toMatchInlineSnapshot(
-        `"0x3fcbf77b28c96f4f2fb5bd2d176ab083a12a5e123adeb0de955d7ee228c9854"`
+        `"0x58d97f7d76e78f44905cc30cb65b91ea49a4b908a76703c54197bca90f81773"`
       );
     });